requirements.txt
streamlit
pandas
plotly
numpy
Pillow
openpyxl